updaed DB clean

+++ server/src/seeds/cleanDb.ts
@@ -1,16 +1,26 @@
-import models from '../models/index.js';
-import db from '../config/connection.js';
+import models from "../models/index.js";
+import db from "../config/connection.js";
 
 export default async (modelName: "Question", collectionName: string) => {
   try {
-    let modelExists = await models[modelName].db.db.listCollections({
-      name: collectionName
-    }).toArray()
+    const model = models[modelName];
+    if (!model || !model.db) {
+      throw new Error(
+        `Model or database connection for "${modelName}" is undefined.`
+      );
+    }
+
+    const modelExists =
+      (await model.db?.db
+        ?.listCollections({
+          name: collectionName,
+        })
+        ?.toArray()) || [];
 
     if (modelExists.length) {
       await db.dropCollection(collectionName);
     }
   } catch (err) {
-    throw err;
+    console.error("Error cleaning database:", err);
   }
-}
+};
